requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.0
requests==2.32.3
streamlit==1.38.0
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
numpy==2.1.1
sentence-transformers==3.0.1
ollama==0.1.9